+++ External/FEXCore/Source/CMakeLists.txt
@@ -311,7 +311,7 @@ endfunction()
 
 function(AddLibrary Name Type)
   add_library(${Name} ${Type} $<TARGET_OBJECTS:${PROJECT_NAME}_object>)
-  target_link_libraries(${Name} vixl dl fmt::fmt xxhash)
+  target_link_libraries(${Name} vixl dl fmt::fmt xxhash tiny-json)
   set_target_properties(${Name} PROPERTIES OUTPUT_NAME FEXCore)
   set_target_properties(${Name} PROPERTIES C_VISIBILITY_PRESET hidden)
   set_target_properties(${Name} PROPERTIES CXX_VISIBILITY_PRESET hidden)
