+++ Source/Common/CMakeLists.txt
@@ -8,6 +8,6 @@ set(SRCS
   StringUtil.cpp)
 
 add_library(${NAME} STATIC ${SRCS})
-target_link_libraries(${NAME} cpp-optparse tiny-json json-maker FEXCore)
+target_link_libraries(${NAME} cpp-optparse json-maker FEXCore)
 target_include_directories(${NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/External/cpp-optparse/)
 target_include_directories(${NAME} PRIVATE ${CMAKE_BINARY_DIR}/generated)
